cmd/link: fix instruction relocations on arm64be
On Aarch64, instructions are always encoded in little-endian.
This means that target.Arch.ByteOrder isn't appropriate when dealing
with relocations that change instructions. Add a special case in the
linker to process these instructions in little-endian instead.

As a consequence, some of the `if target.IsBigEndian()` special-cases in
arm64/asm.go aren't relevant anymore.

This patch fixes the linking of the first call instruction in a arm64be
program. TODO: example.